+++ bootstrap/helpers.php
@@ -4,7 +4,9 @@ namespace Busarm\PhpMini\Helpers;
 
 use Busarm\PhpMini\Async;
 use Busarm\PhpMini\Errors\SystemError;
+use Busarm\PhpMini\Interfaces\Runnable;
 use Busarm\PhpMini\Promise;
+use Busarm\PhpMini\Tasks\CallableTask;
 use Busarm\PhpMini\Tasks\Task;
 use Closure;
 use Generator;
@@ -528,11 +530,11 @@ function concurrent(array $tasks, $wait = false): Generator
  * Listen to event
  * 
  * @param string $event
- * @param callable|class-string<Task> $listner
+ * @param callable|class-string<Runnable> $listner
  */
 function listen(string $event, callable|string $listner)
 {
-    app()->eventManager->addEventListner($event, $listner);
+    app()->eventHandler?->listen($event, $listner);
 }
 
 /**
@@ -543,9 +545,19 @@ function listen(string $event, callable|string $listner)
  */
 function dispatch(string $event, array $data = [])
 {
-    app()->eventManager->dispatchEvent($event, $data);
+    app()->eventHandler?->dispatch($event, $data);
 }
 
+/**
+ * Queue task
+ * 
+ * @param Task|callable $task
+ */
+function enqueue(Task|callable $task)
+{
+    $task = $task instanceof Task ? $task : new CallableTask(Closure::fromCallable($task));
+    app()->queueHandler?->enqueue($task);
+}
 
 /**
  * Wrap data to be serialized
@@ -553,7 +565,7 @@ function dispatch(string $event, array $data = [])
  * @param mixed $data
  * @return mixed
  */
-function wrapSerializable($data)
+function wrap_serializable($data)
 {
     if ($data instanceof Closure) {
         $data = new SerializableClosure($data);
@@ -561,7 +573,7 @@ function wrapSerializable($data)
         $data = new SerializableClosure(Closure::fromCallable($data));
     } else if (is_array($data)) {
         $data = array_map(function ($value) {
-            return wrapSerializable($value);
+            return wrap_serializable($value);
         }, $data);
     } else if (is_object($data)) {
         $reflection = new ReflectionObject($data);
@@ -576,7 +588,7 @@ function wrapSerializable($data)
                     ) {
                         $value = $prop->getValue($data);
                         if (isset($value)) {
-                            $prop->setValue($data, wrapSerializable($value));
+                            $prop->setValue($data, wrap_serializable($value));
                         }
                     }
                 }
@@ -592,13 +604,13 @@ function wrapSerializable($data)
  * @param mixed $data
  * @return mixed
  */
-function unwrapSerializable($data)
+function unwrap_serializable($data)
 {
     if ($data instanceof SerializableClosure) {
         $data = $data->getClosure();
     } else if (is_array($data)) {
         $data = array_map(function ($value) {
-            return unwrapSerializable($value);
+            return unwrap_serializable($value);
         }, $data);
     } else if (is_object($data)) {
         $reflection = new ReflectionObject($data);
@@ -613,7 +625,7 @@ function unwrapSerializable($data)
                     ) {
                         $value = $prop->getValue($data);
                         if (isset($value)) {
-                            $prop->setValue($data, unwrapSerializable($value));
+                            $prop->setValue($data, unwrap_serializable($value));
                         }
                     }
                 }
@@ -631,7 +643,7 @@ function unwrapSerializable($data)
  */
 function serialize($data)
 {
-    return \serialize(wrapSerializable($data));
+    return \serialize(wrap_serializable($data));
 }
 
 /**
@@ -643,7 +655,7 @@ function serialize($data)
  */
 function unserialize($data, array $options = [])
 {
-    return unwrapSerializable(\unserialize($data, $options));
+    return unwrap_serializable(\unserialize($data, $options));
 }
 
 /**