+++ src/Errors/QueueError.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Busarm\PhpMini\Errors;
+
+/**
+ * PHP Mini Framework
+ *
+ * @copyright busarm.com
+ * @license https://github.com/Busarm/php-mini/blob/master/LICENSE (MIT License)
+ * @codeCoverageIgnore
+ */
+class QueueError extends SystemError
+{
+}