+++ src/Promise.php
@@ -6,6 +6,7 @@ use Busarm\PhpMini\Interfaces\Promise\PromiseCatch;
 use Busarm\PhpMini\Interfaces\Promise\PromiseFinal;
 use Busarm\PhpMini\Interfaces\Promise\PromiseThen;
 use Busarm\PhpMini\Tasks\CallableTask;
+use Busarm\PhpMini\Tasks\Task;
 use Closure;
 use Fiber;
 use Generator;
@@ -23,7 +24,7 @@ use function Busarm\PhpMini\Helpers\report;
  * @inheritDoc
  * 
  * ## CAUTION ! ##
- * Be careful when using parameters with the `callable`. 
+ * Be careful when using parameters with a callable `task`. 
  * Parameters will also be serialized before sending to the task worker.
  * Hence, ensure that the parameters are simple. 
  * 
@@ -63,12 +64,12 @@ class Promise implements PromiseThen, PromiseCatch, PromiseFinal
     private ?Closure $_finallyFn = null;
 
     /**
-     * @param callable(array $params):T $callable 
-     * @param array $params
+     * @param Task|callable(array $params):T $task 
      */
-    public function __construct(callable $callable, ...$params)
+    public function __construct(Task|callable $task)
     {
-        $task = new CallableTask(Closure::fromCallable($callable), $params);
+        $task = $task instanceof Task ? $task : new CallableTask(Closure::fromCallable($task));
+
         $this->_id = $task->getName();
         $this->_fiber = Async::withFiberWorker($this->_id, strval($task->getRequest(false)), true);
     }
@@ -154,28 +155,13 @@ class Promise implements PromiseThen, PromiseCatch, PromiseFinal
     }
 
     /**
-     * @param mixed $value
-     * @return mixed
-     */
-    public static function suspend($value = null)
-    {
-        return Fiber::suspend($value);
-    }
-
-    /**
-     * Run tasks concurrently
+     * Resolve list of promises
      * 
-     * @param callable[] $tasks List of Task instance to run
+     * @param self[] $promises List of Promises
      * @return Generator
      */
-    public static function all(array $tasks): Generator
+    public static function all(array $promises): Generator
     {
-        /** @var self[] */
-        $promises = [];
-        foreach ($tasks as $key => $task) {
-            $promises[$key] = new self($task);
-        }
-        ksort($promises);
         foreach ($promises as $key => $promise) {
             yield $key => $promise->wait();
         }